Change title of "Active" column
In the admin panel, in the students section, the first
column is the "Active" column. In this case, active
means that the student doesn't have any pending invoices.

To make this more self explanatory, instead of using the
word "Active" we switched to the spanish term "Al día"
which means that the student doesn't have any pending
invoices.